// src/types/image.ts
export interface ImageAnalyseInfos {
  average_distance_max_delta_percent: number;
  color: number[];
  number_of_pixel_with_delta_20_10_percent_max_delta: number;
  number_of_pixel_with_delta_30_20_percent_max_delta: number;
  number_of_pixel_with_delta_40_30_percent_max_delta: number;
  number_of_pixel_with_delta_50_40_percent_max_delta: number;
  number_of_pixel_with_delta_60_50_percent_max_delta: number;
  number_of_pixel_with_delta_70_60_percent_max_delta: number;
  number_of_pixel_with_delta_80_70_percent_max_delta: number;
  number_of_pixel_with_delta_90_80_percent_max_delta: number;
  number_of_pixel_with_delta_upper_90_percent_max_delta: number;
  number_of_pixel_with_delta_minus_or_equal_10_percent_max_delta: number;
  size: number;
}